requirements.txt
annotated-types==0.6.0
anyio==3.7.1
apeye==1.4.1
apeye-core==1.1.4
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
coloredlogs==15.0.1
conda-pack==0.7.1
dimits==0.0.21a0
domdf-python-tools==3.7.0
espeak-phonemizer==1.3.1
fastapi==0.104.0
flatbuffers==23.5.26
h11==0.14.0
humanfriendly==10.0
idna==3.4
mpmath==1.3.0
natsort==8.4.0
numpy==1.26.1
onnxruntime==1.16.1
packaging==23.2
piper-phonemize==1.0.0
piper-tts==1.1.0
platformdirs==3.11.0
protobuf==4.24.4
pydantic==2.4.2
pydantic_core==2.10.1
pypi-json==0.4.0
requests==2.31.0
shortuuid==1.0.11
sniffio==1.3.0
sympy==1.12
tqdm==4.64.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
